backend/requirements.txt
bcrypt==4.1.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3